feat(getmoonbeam-app): #31 - offer front-end implementation draft 1.
This implements the initial loading of offers for the Store. It is pending individual
store details and individual store webview front-end integration.

MOONBM-31

+++ packages/moonbeam-models/src/common/GraphqlExports.ts
@@ -656,7 +656,7 @@ export type Offer = {
   extOfferId?: Maybe<Scalars['ID']>;
   id?: Maybe<Scalars['ID']>;
   offerState?: Maybe<OfferState>;
-  purchaseAmount?: Maybe<Scalars['Int']>;
+  purchaseAmount?: Maybe<Scalars['Float']>;
   purchaseFrequency?: Maybe<Scalars['Int']>;
   qualifier?: Maybe<Scalars['String']>;
   reach?: Maybe<OfferReach>;
